#43 Move literals, term parser outside CqlParser
